clippy fixes

+++ src/cli.rs
@@ -1234,7 +1234,7 @@ pub fn run(opt: Varlociraptor) -> Result<()> {
                 &sample,
                 coding_genome_size as u64,
                 mode,
-                cutoff as f64,
+                cutoff,
             )?,
             EstimateKind::AlignmentProperties {
                 reference,
