+++ src/variants/types/replacement.rs
@@ -77,7 +77,7 @@ impl<R: Realigner> Realignable for Replacement<R> {
         _: &genome::Interval,
         ref_window: usize,
     ) -> Result<Vec<Box<dyn RefBaseVariantEmission>>> {
-        let repl_alt_len = self.replacement.len() as usize;
+        let repl_alt_len = self.replacement.len();
         let repl_ref_len = self.ref_len();
 
         let start = self.locus().range().start as usize;
